[TESTED WITH CLASSIC PARAMETERS] Beeman works but does not appear to converge more rapidly than other methods... Need further tests (Not predictor corrector but with correction of the professor in which the a(t + dt) is calculated using v(t) and not v(t+dt).

+++ models/src/main/java/ar/edu/itba/ss/cushioned_oscilator/models/Vector2D.java
@@ -42,7 +42,7 @@ public class Vector2D {
     return x;
   }
 
-  public void setX(final int x) {
+  public void setX(final double x) {
     this.x = x;
   }
 
@@ -50,7 +50,7 @@ public class Vector2D {
     return y;
   }
 
-  public void setY(final int y) {
+  public void setY(final double y) {
     this.y = y;
   }
